+++ src/backend/impls/rng_cpu.c
@@ -13,6 +13,7 @@
 #include <stdint.h>
 
 #include "backend.h"
+#include "backend_private.h"
 #include "utils.h"
 
 /*
@@ -63,6 +64,16 @@ static inline void reduce_uniform_n(int64_t* tgt, int n_bits)
 	*tgt     = (int64_t)((uint64_t)(*tgt) << shft) >> shft;
 }
 
+/**
+ * @brief Generates a uniformly sampled random number in [-2^(nb_bits-1), 2^(nb_bits-1))
+ *
+ * @param module  The backend object
+ * @param result  The resulting uniformly sampled integer
+ * @param nb_bits The number of bits of the result
+ *
+ * @retval -1 if an error occurs.
+ * @retval 0 otherwise.
+ */
 int ref_rand_uniform_pow2(const PvdaBackend* module, int64_t* result, uint64_t nb_bits)
 {
 	// As result points to an uint64_t  nb_bits shall not exceed its size
@@ -84,6 +95,19 @@ int ref_rand_uniform_pow2(const PvdaBackend* module, int64_t* result, uint64_t n
 	return 0;
 }
 
+/**
+ * @brief Generates a uniformly sampled random number in [limit_down, limit_up] via
+ * power-of-2 sampling and resampling if out-of-bounds
+ *
+ * @param module      The backend object
+ * @param result      The resulting uniformly sampled integer
+ * @param limit_down  The lower bound of the uniform sample
+ * @param limit_up    The upper bound of the uniform sample
+ *
+ *
+ * @retval -1 if an error occurs.
+ * @retval 0 otherwise.
+ */
 int ref_rand_uniform(const PvdaBackend* module, int64_t* result, int64_t limit_down, int64_t limit_up)
 {
 	uint64_t max_delta = (uint64_t)limit_up - (uint64_t)limit_down;
@@ -104,6 +128,19 @@ int ref_rand_uniform(const PvdaBackend* module, int64_t* result, int64_t limit_d
 	return 0;
 }
 
+/**
+ * @brief Generates a uniformly random vector of numbers
+ *
+ * Coefficients are uniformly sampled in range [-2^(nb_bits-1), 2^(nb_bits-1))
+ *
+ * @param module  The backend object
+ * @param res     The result uniformly drawn vector of numbers
+ * @param nn      Number of coeffients in the polynomial (eq. degree of the cyclotomial poly)
+ * @param nb_bits Number of randomness bits per coefficient.
+ *
+ * @retval -1 if an error occurs
+ * @retval 0 otherwise.
+ */
 int ref_rand_uniform_pow2_vec(const PvdaBackend* module, int64_t* res, uint64_t n, uint64_t nb_bits)
 {
 	CHECK_CALL(cpu_read_rand((uint64_t*)res, sizeof(int64_t) * n), "rng error");
@@ -116,6 +153,18 @@ cleanup:
 	return -1;
 }
 
+/**
+ * @brief Generates a uniformly random binary vector of numbers
+ *
+ * Coefficients are uniformly sampled in {0, 1}
+ *
+ * @param module  The backend object
+ * @param res     The result uniformly drawn \ZnX polynomial.
+ * @param nn      Number of coeffients in the polynomial (eq. degree of the cyclotomial poly)
+ *
+ * @retval -1 if an error occurs
+ * @retval 0 otherwise.
+ */
 int ref_rand_uniform_binary_vec(const PvdaBackend* module, uint64_t* res, uint64_t n)
 {
 	CHECK_CALL(cpu_read_rand(res, sizeof(uint64_t) * n), "rng error");
@@ -126,4 +175,12 @@ int ref_rand_uniform_binary_vec(const PvdaBackend* module, uint64_t* res, uint64
 	return 0;
 cleanup:
 	return -1;
+}
+
+void pvda_fill_ref_rng(struct pvda_virtual_table* vt)
+{
+	vt->pvda_rand_uniform_pow2       = ref_rand_uniform_pow2;
+	vt->pvda_rand_uniform_pow2_vec   = ref_rand_uniform_pow2_vec;
+	vt->pvda_rand_uniform_binary_vec = ref_rand_uniform_binary_vec;
+	vt->pvda_rand_uniform            = ref_rand_uniform;
 }
